+++ src/pages/me.tsx
@@ -1,5 +1,5 @@
 import { Router } from "@reach/router"
-import React from "react"
+import React, { FC } from "react"
 import Details from "../components/Auth/Details"
 import Login from "../components/Auth/Login"
 import PrivateRoute from "../components/Auth/PrivateRoute"
@@ -7,7 +7,7 @@ import SignUp from "../components/Auth/Register"
 import Reset from "../components/Auth/Reset"
 import Layout from "../components/layout"
 
-const Me = () => {
+const Me: FC = () => {
   return (
     <Layout>
       <Router>
